requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
pandas>=2.1.0
tqdm>=4.66.0
matplotlib>=3.8.0
scikit-learn>=1.4.0
pyyaml>=6.0.1
pytest>=8.0.0
